Changed to use getX
