+++ lib/blip_ds.dart
@@ -1,7 +1,7 @@
 library blip_ds;
 
 export 'src/components/chat/ds_message_bubble.dart' show DSMessageBubble;
-export 'src/components/chat/ds_audio_message_bubble.dart'
+export 'src/components/chat/audio/ds_audio_message_bubble.dart'
     show DSAudioMessageBubble;
 export 'src/enums/ds_border_radius.enum.dart' show DSBorderRadius;
 export 'src/enums/ds_align.enum.dart' show DSAlign;
