+++ pubspec.yaml
@@ -14,6 +14,7 @@ dependencies:
   just_audio: ^0.9.27
   rxdart: ^0.27.4
   flutter_spinkit: ^5.1.0
+  get: ^4.6.5
 
 dev_dependencies:
   flutter_test:
