+++ sample/pubspec.yaml
@@ -34,6 +34,7 @@ dependencies:
   just_audio: ^0.9.27
   rxdart: ^0.27.4
   flutter_spinkit: ^5.1.0
+  get: ^4.6.5
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
